Change Theme and ViewModel initialization in BaseView, Refactoring Routes and Themes directory
